feat: Add a wrapper lib around our vector DB to simplify the code and simplify future updates
