+++ src/workers/indexMarkdown.ts
@@ -1,16 +1,15 @@
-import { ChromaClient } from 'chromadb'
-import chromadb from '../config/chromadb'
 import { DiscordWorker, DiscordJob } from '../lib/DiscordWorker'
+import { vectorDB } from '../lib/vectordb'
 
 class JobData extends DiscordJob {}
 
 const indexMarkdown = new DiscordWorker(
   'indexMarkdown',
   async (job: JobData) => {
-    const client = new ChromaClient(chromadb)
     const { url } = job.data
     const childrenValues = await job.getChildrenEntries()
-    const { markdown } = childrenValues
+    const { markdown }: { markdown: string } = childrenValues
+
     const paragraphs = markdown
       .split('\n###')
       .map((p) => p.trim())
@@ -18,25 +17,9 @@ const indexMarkdown = new DiscordWorker(
 
     await job.sendMessage(`🤖 Sparar i vektordatabas...`)
     job.log('Indexing ' + paragraphs.length + ' paragraphs from url: ' + url)
+
     try {
-      const collection = await client.getOrCreateCollection({
-        name: 'emission_reports',
-        embeddingFunction: job.embedder,
-      })
-      job.log('Indexing ' + paragraphs.length + ' paragraphs...')
-      const ids = paragraphs.map((p, i) => job.data.url + '#' + i)
-      const metadatas = paragraphs.map((p, i) => ({
-        source: url,
-        markdown,
-        type: 'company_sustainability_report', // this is our own type to be able to filter in the future if needed
-        parsed: new Date().toISOString(),
-        page: i,
-      }))
-      await collection.add({
-        ids,
-        metadatas,
-        documents: paragraphs,
-      })
+      await vectorDB.addReport(url, markdown, paragraphs)
       job.editMessage(`✅ Sparad i vektordatabasen`)
       job.log('Done!')
